Prepare for test deployment

+++ controllers/products.js
@@ -2,9 +2,7 @@ const _ = require("lodash");
 const { Product, validate, validateOnUpdate } = require("../models/Product");
 const { ProductCategory } = require("../models/ProductCategory");
 const { Discount } = require("../models/Discount");
-const validateObjectId = require("../utils/validateObjectId");
 const uploader = require("../utils/uploader");
-const routeErrorHandler = require("../middleware/routeErrorHandler");
 
 const createProduct = async (req, res) => {
   const { error } = validate(req.body);
@@ -65,19 +63,24 @@ const updateProduct = async (req, res) => {
   let product = await Product.findById(req.params.id);
   if (!product) return res.status(404).send("Product not found.");
 
-  if (product?.image?.url) {
-    if (product.image.url !== req.body.imageUri)
-      await uploader.deleteFile(product.image.public_id);
+  if (product?.image?.url && req.body.imageUri) {
+    if (product.image.url != req.body.imageUri)
+      try {
+        await uploader.deleteFile(product.image.public_id);
+      } catch (err) {
+        return res.status(400).send("Oops... could not update product.");
+      }
   }
 
   product.name = req.body.name;
   product.desc = req.body.desc;
   product.category = req.body.categoryId;
   product.price = req.body.price;
-  product.image.url = req.body.imageUri;
   product.image.public_id = req.body?.imagePublicId;
   product.status = req.body.status;
   product.discount = req.body?.discountId;
+
+  if (req.body.imageUri) product.image.url = req.body.imageUri;
 
   await product.save();
   return res.send(product);
@@ -87,8 +90,13 @@ const deleteProduct = async (req, res) => {
   let product = await Product.findById(req.params.id);
   if (!product) return res.status(404).send("Product not found.");
 
-  if (product?.image?.public_id)
-    await uploader.deleteFile(product?.image?.public_id);
+  if (product?.image?.public_id) {
+    try {
+      await uploader.deleteFile(product?.image?.public_id);
+    } catch (error) {
+      return res.status(500).send("Oops. Could not delete product.");
+    }
+  }
 
   product.remove();
